chore: refactor to more readable file structure [still needs some work]
