+++ packages/web/src/pages/_protected/groups/$slug/settings.page.tsx
@@ -1,15 +1,15 @@
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useDeleteGroup, useGetGroupBySlug } from "../@data";
 import { GroupBody, SecondaryRow } from "./layout";
 import { SubHeading } from "@/components/prose";
 import { useWithConfirmation } from "@/components/with-confirmation-dialog";
 import { Button } from "@/components/ui/button";
 import { States } from "./layout";
+import { useDeleteGroup, useGroupBySlug } from "../../@data/groups";
 
 function SettingsRoute() {
   const navigate = useNavigate();
   const params = Route.useParams();
-  const { data, status } = useGetGroupBySlug(params.slug);
+  const { data, status } = useGroupBySlug(params.slug);
   const deleteGroup = useDeleteGroup();
 
   if (status === "not-found") return <States.NotFound title={params.slug} />;
