+++ packages/web/src/pages/_protected/layout.tsx
@@ -20,16 +20,17 @@ import {
   useMatches,
 } from "@tanstack/react-router";
 import { Fragment } from "react/jsx-runtime";
-import { GlobalCommandBar } from "./@command-bar";
-import { ZeroProvider } from "@/lib/zero.provider";
-import { authenticationQueryOptions, AuthProvider } from "@/lib/auth.provider";
+import { GlobalCommandBar } from "./@command-bar.dialog";
+import { ZeroProvider } from "@/lib/zero/zero-provider";
+import { authenticationQueryOptions } from "@/lib/authentication";
+import { AuthProvider } from "@/lib/authentication/auth-provider";
 import { PropsWithChildren } from "react";
-import { CreateExpenseDialog } from "./@create-expense";
-import { CreateGroupDialog } from "./groups/@create-group";
+import { CreateExpenseDialog } from "./@create-expense.dialog";
+import { CreateGroupDialog } from "./groups/@create-group.dialog";
 import { Toaster } from "@/components/ui/sonner";
-import { SearchRouteSchema as GlobalSearchParams } from "./@command-bar";
-import { SearchRouteSchema as CreateExpenseSearchParams } from "./@create-expense";
-import { SearchRouteSchema as CreateGroupSearchParams } from "./groups/@create-group";
+import { SearchRouteSchema as GlobalSearchParams } from "./@command-bar.dialog";
+import { SearchRouteSchema as CreateExpenseSearchParams } from "./@create-expense.dialog";
+import { SearchRouteSchema as CreateGroupSearchParams } from "./groups/@create-group.dialog";
 import * as v from "valibot";
 
 function getCookie(name: string, fallback: string) {
