+++ packages/web/src/router.tsx
@@ -42,6 +42,7 @@ declare module "@tanstack/react-router" {
 }
 
 declare module "@tanstack/react-table" {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   interface TableMeta<TData extends RowData> {
     expand: (id: string) => void;
     updateTitle: (id: string) => void;
